requirements.txt
alembic==1.13.1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==23.2.0
Babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.7.0
Brotli==1.1.0
bs4==0.0.2
cattrs==23.2.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
cssselect2==0.7.0
Deprecated==1.2.14
dnspython==1.16.0
et-xmlfile==1.1.0
Flask==3.0.2
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-FlatPages==0.8.2
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-PyMongo==2.3.0
Flask-Sitemap==0.4.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Toastr==0.5.8
flatbuffers==24.3.25
fonttools==4.50.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
idna==3.6
importlib_resources==6.4.4
itsdangerous==2.1.2
Jinja2==3.1.3
limits==3.13.0
lxml==5.2.1
mailersend==0.5.6
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient==2.2.4
numpy==1.26.4
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.4
openpyxl==3.1.2
ordered-set==4.1.0
oscrypto==1.3.0
packaging==23.2
pandas==2.2.2
pdfkit==1.0.0
pillow==10.3.0
platformdirs==4.2.2
pycparser==2.21
pydyf==0.9.0
Pygments==2.18.0
pyHanko==0.24.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pymongo==3.11.0
PyMySQL==1.1.1
pypdf==4.2.0
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
railway==0.0.4
reportlab==4.0.9
requests==2.31.0
requests-cache==1.2.0
retry-requests==2.0.0
rich==13.8.0
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.28
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
url-normalize==1.4.3
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
zope.event==5.0
zope.interface==6.1
zopfli==0.2.3
